moved permutation list to object package
