+++ service/spec/permutation.go
@@ -1,5 +1,9 @@
 package spec
 
+import (
+	objectspec "github.com/xh3b4sd/anna/object/spec"
+)
+
 // Permutation creates permutations of arbitrary lists as configured.
 type Permutation interface {
 	// PermuteBy permutes the configured values by applying the given delta to
@@ -28,27 +32,5 @@ type Permutation interface {
 	//
 	//         []interface{"a", "a"}
 	//
-	PermuteBy(list PermutationList, delta int) error
-}
-
-// TODO this should probably be an object and not inside the service package
-
-// PermutationList is supposed to be permuted by a permutation service.
-type PermutationList interface {
-	// GetIndizes returns the list's current indizes.
-	GetIndizes() []int
-
-	// GetMaxGrowth returns the list's configured growth limit. The growth limit
-	// is used to prevent infinite permutations. E.g. MaxGrowth set to 4 will not
-	// permute up to a list of 5 raw values.
-	GetMaxGrowth() int
-
-	// GetPermutedValues returns the list's permuted values.
-	GetPermutedValues() []interface{}
-
-	// GetRawValues returns the list's unpermuted raw values.
-	GetRawValues() []interface{}
-
-	// SetIndizes sets the given indizes of the current permutation list.
-	SetIndizes(indizes []int)
+	PermuteBy(list objectspec.PermutationList, delta int) error
 }
